Simple Hex Server interface
