+++ server/HexDLC.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "GAlgorithm.h"
+
+// Thread safe HexDLC code
+class HexDLC {
+public:
+	HexDLC();
+	void run(); // Run this on different thread
+	Member getChampion();
+private:
+	Population mPopulation;
+	bool mDLCRunning = true;
+
+	bool mChampionRequested = false;
+	Member* mChampion = nullptr;
+};
