+++ server/MyRandom.cpp
